Move birthday config DTOs to peopleplanner and validate the update request
