+++ backend/src/main/java/com/skapp/community/peopleplanner/service/PeopleConfigService.java
@@ -1,7 +1,7 @@
 package com.skapp.community.peopleplanner.service;
 
-import com.skapp.community.common.payload.request.BirthdayNotificationConfigRequestDto;
 import com.skapp.community.common.payload.response.ResponseEntityDto;
+import com.skapp.community.peopleplanner.payload.request.BirthdayNotificationConfigRequestDto;
 
 public interface PeopleConfigService {
